Do not register hits on key hold; Fix issues with scores.

+++ source/Graphics/pDrawable.cpp
@@ -83,7 +83,7 @@ bool pDrawable::InBounds(int32_t x, int32_t y)
 	}
 }
 
-void pDrawable::Kill(int32_t time)
+void pDrawable::Kill(long time)
 {
 	Transform(TR_KILL, time, time, 0, 0);
 }
@@ -103,24 +103,24 @@ void pDrawable::Transform(Transformation* transform)
 	mTransformations.push_back(transform);
 }
 
-void pDrawable::Transform(TransformType type, int32_t starttime, int32_t endtime, int32_t startvalue, int32_t endvalue)
+void pDrawable::Transform(TransformType type, long starttime, long endtime, int32_t startvalue, int32_t endvalue)
 {
 	Transform(new Transformation(type, starttime, endtime, startvalue, endvalue));
 }
 
-void pDrawable::Scale(int32_t starttime, int32_t endtime, float start, float end)
+void pDrawable::Scale(long starttime, long endtime, float start, float end)
 {
 	Transform(TR_SCALEX, starttime, endtime, mOrigWidth*start, mOrigWidth*end);
 	Transform(TR_SCALEY, starttime, endtime, mOrigHeight*start, mOrigHeight*end);
 }
 
-void pDrawable::Move(int32_t starttime, int32_t endtime, int32_t startx, int32_t starty, int32_t endx, int32_t endy)
+void pDrawable::Move(long starttime, long endtime, int32_t startx, int32_t starty, int32_t endx, int32_t endy)
 {
 	Transform(TR_MOVEX, starttime, endtime, startx, endx);
 	Transform(TR_MOVEY, starttime, endtime, starty, endy);
 }
 
-void pDrawable::Move(int32_t starttime, int32_t endtime, int32_t endx, int32_t endy)
+void pDrawable::Move(long starttime, long endtime, int32_t endx, int32_t endy)
 {
 	Move(starttime, endtime, X, Y, endx, endy);
 }
@@ -131,7 +131,7 @@ void pDrawable::Move(int32_t endx, int32_t endy)
 	Y = endy;
 }
 
-void pDrawable::Rotate(int32_t starttime, int32_t endtime, int32_t starta, int32_t enda)
+void pDrawable::Rotate(long starttime, long endtime, int32_t starta, int32_t enda)
 {
 	Transform(TR_ROTATE, starttime, endtime, starta, enda);
 }
@@ -146,12 +146,12 @@ void pDrawable::Show()
 	Show(GameClock::Clock().Time());
 }
 
-void pDrawable::Show(int32_t time)
+void pDrawable::Show(long time)
 {
 	Transform(TR_FADE, time, time, 255, 255);
 }
 
-void pDrawable::Show(int32_t starttime, int32_t endtime)
+void pDrawable::Show(long starttime, long endtime)
 {
 	Transform(TR_FADE, starttime, endtime, 0, 255);
 }
@@ -161,12 +161,12 @@ void pDrawable::Hide()
 	Hide(GameClock::Clock().Time());
 }
 
-void pDrawable::Hide(int32_t time)
+void pDrawable::Hide(long time)
 {
 	Transform(TR_FADE, time, time, 0, 0);
 }
 
-void pDrawable::Hide(int32_t starttime, int32_t endtime)
+void pDrawable::Hide(long starttime, long endtime)
 {
 	Transform(TR_FADE, starttime, endtime, 255, 0);
 }
